Moved preview/vuln tests to their own packages.
