+++ cmd/fleetctl/integrationtest/vuln/vulnerability_data_stream_test.go
@@ -1,4 +1,4 @@
-package fleetctl
+package vuln
 
 import (
 	"fmt"
@@ -6,6 +6,7 @@ import (
 	"path"
 	"testing"
 
+	"github.com/fleetdm/fleet/v4/cmd/fleetctl/fleetctl"
 	"github.com/fleetdm/fleet/v4/pkg/nettest"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -14,7 +15,7 @@ import (
 func TestIntegrationsVulnerabilityDataStream(t *testing.T) {
 	nettest.Run(t)
 
-	RunAppCheckErr(t, []string{"vulnerability-data-stream"}, "No directory provided")
+	fleetctl.RunAppCheckErr(t, []string{"vulnerability-data-stream"}, "No directory provided")
 
 	vulnPath := t.TempDir()
 	expectedOutput := `[-] Downloading CPE database... Done
@@ -34,7 +35,7 @@ func TestIntegrationsVulnerabilityDataStream(t *testing.T) {
 
 	var actualOutput string
 	err := nettest.RunWithNetRetry(t, func() error {
-		w, err := RunAppNoChecks([]string{"vulnerability-data-stream", "--dir", vulnPath})
+		w, err := fleetctl.RunAppNoChecks([]string{"vulnerability-data-stream", "--dir", vulnPath})
 		actualOutput = w.String()
 		return err
 	})
